Tricky test 3

+++ main-3-5.cpp
@@ -5,7 +5,7 @@
 extern double sum_even(double array[], int n);
 
 int main() {
-    int array[5] = {4,5,6,7,8};
+    double array[5] = {4,5,6,7,8};
     std::cout << "The sum of elements in even positions is: " << sum_even(array, 5) << std::endl;
     return 0;
 }
